Is the math right on this?
I was getting huge numbers in my comparisons and looking at the code, it seems like we should divide the avg_sims by the number of averages in that array in order to get the "total_avg"? Otherwise we are just adding all the averages together and getting a number that doesn't really tell us anything valuable. Am I wrong?